pyproject.toml
[tool.poetry]
name = "todo-fastapi-grpc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.65.1"
grpcio-tools = "^1.65.1"
uvicorn = "^0.30.3"
fastapi = "^0.111.1"
piccolo = {extras = ["sqlite"], version = "^1.14.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
